requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
promptlayer>=1.0.0
openai>=1.75.0
anthropic>=0.39.0
httpx>=0.27.0
tavily-python>=0.3.0
supabase>=2.0.0
google-generativeai>=0.8.0
google-genai>=0.3.0
google-api-python-client>=2.100.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
jinja2>=3.1.0
langgraph>=0.4.0
langgraph-checkpoint-sqlite>=2.0.0